internal/controld: check if ipv4 is available before connect to API
Updates #53
